// src/common/proto/project.pb.ts
export enum LevelType {
  LOW = 0,
  HIGH = 1,
  UNRECOGNIZED = -1,
}

export enum UnityType {
  VB = 0,
  M2 = 1,
  UNID = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface PossibleSkills {
  skillId: string;
  requiredSkillLevel: number;
}

/**
 * TaskCreate
 * Request
 */
export interface TaskCreateRequest {
  category: string;
  activity: string;
  noiseLevel: LevelType;
  dirtLevel: LevelType;
  description: string;
  unity: UnityType;
  possibleSkills: PossibleSkills[];
}

/** Response */
export interface TaskCreateResponse {
  status: number;
  error: string;
  data: TaskData | undefined;
}

/**
 * TaskFindAllRequest
 * Request
 */
export interface TaskFindAllRequest {}

/** Response */
export interface TaskFindAllResponse {
  status: number;
  error: string;
  data: TaskData[];
}

/**
 * TaskFindOne
 * Request
 */
export interface TaskFindOneRequest {
  id: string;
}

/** Response */
export interface TaskFindOneResponse {
  status: number;
  error: string;
  data: TaskData | undefined;
}

/**
 * TaskUpdate
 * Request
 */
export interface TaskUpdateRequest {
  id: string;
  data: TaskUpdateData | undefined;
}

/** Response */
export interface TaskUpdateResponse {
  status: number;
  error: string;
  data: TaskData | undefined;
}

export interface TaskUpdateData {
  category?: string | undefined;
  activity?: string | undefined;
  noiseLevel?: string | undefined;
  dirtLevel?: string | undefined;
}

/**
 * TaskRemove
 * Request
 */
export interface TaskRemoveRequest {
  id: string;
}